+++ src/frontend/packages/lib_video/src/api/createThumbnail/index.ts
@@ -1,16 +1,25 @@
-import { useJwt, API_ENDPOINT, modelName, Thumbnail } from 'lib-components';
+import {
+  fetchWrapper,
+  useJwt,
+  API_ENDPOINT,
+  modelName,
+  Thumbnail,
+} from 'lib-components';
 
 /**
  * Create a new thumbnail record for a language-mode combination.
  */
 export const createThumbnail = async () => {
-  const response = await fetch(`${API_ENDPOINT}/${modelName.THUMBNAILS}/`, {
-    headers: {
-      Authorization: `Bearer ${useJwt.getState().jwt ?? ''}`,
-      'Content-Type': 'application/json',
+  const response = await fetchWrapper(
+    `${API_ENDPOINT}/${modelName.THUMBNAILS}/`,
+    {
+      headers: {
+        Authorization: `Bearer ${useJwt.getState().jwt ?? ''}`,
+        'Content-Type': 'application/json',
+      },
+      method: 'POST',
     },
-    method: 'POST',
-  });
+  );
 
   if (!response.ok) {
     throw new Error('Failed to create a new thumbnail.');